glaucoscan-API/requirements.txt
tensorflow>=2.8.0
keras>=2.8.0
matplotlib>=3.5.1
numpy>=1.21.5
pandas>=1.4.2
fastapi>=0.75.0
uvicorn>=0.17.6
python-multipart>=0.0.5
pillow>=9.1.0
pydantic>=1.9.0
scikit-learn>=1.0.2
tqdm>=4.64.0
pytest>=7.1.1
requests>=2.27.1
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
imbalanced-learn>=0.9.0
joblib>=1.1.0
loguru
pyyaml
opencv-python
